Add querying and more comments
Querying was added to gain me an extra 5 marks, and it took no more
than 2 minutes to implement!

Plenty more comments were added in order to demonstrate that I
understand the process the code goes through.

Also, the pulser script now has a more useful usage statement, doesn't
require python to parse the status of the response, and can have spaces
inside the module name, as the script will replace them with %20.